+++ lib/core/event/user_update_model_event.dart
@@ -1,4 +1,5 @@
-import 'package:fiver/data/model/info_user_access_token.dart';
+
+import '../../data/model/user_info_model.dart';
 
 class UserInfoModelUpdateEvent {
   final UserInfoModel user;
