+++ lib/data/model/info_user_access_token.dart
@@ -1,3 +1,5 @@
+import 'user_info_model.dart';
+
 class InfoUserAccessTokenModel {
   UserInfoModel? userInfo;
   String? accessToken;
@@ -24,28 +26,6 @@ class InfoUserAccessTokenModel {
     data['access_token'] = accessToken;
     data['token_type'] = tokenType;
     data['expires_at'] = expiresAt;
-    return data;
-  }
-}
-
-class UserInfoModel {
-  String? fullName;
-  String? email;
-  String? avatar;
-
-  UserInfoModel({this.fullName, this.email, this.avatar});
-
-  UserInfoModel.fromJson(Map<String, dynamic> json) {
-    fullName = json['full_name'];
-    email = json['email'];
-    avatar = json['avatar'];
-  }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = <String, dynamic>{};
-    data['full_name'] = fullName;
-    data['email'] = email;
-    data['avatar'] = avatar;
     return data;
   }
 }
